Renaming covarianceOsc to ParameterHandlerOsc
