+++ covariance/SystematicHandlerGeneric.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // MaCh3 includes
-#include "covariance/covarianceBase.h"
+#include "covariance/ParameterHandlerBase.h"
 #include "samplePDF/Structs.h"
 
 /// @brief Class responsible for handling of systematic error parameters with different types defined in the config. Like spline, normalisation parameters etc.
